+++ src/Application/Import/UploadedFile.php
@@ -1,0 +1,3 @@
+<?php
+namespace EventFlow\Application\Import;
+final readonly class UploadedFile{public function __construct(public string$clientFilename,public string$temporaryPath,public int$reportedSize,public int$errorCode){}}